+++ src/browser-names.js
@@ -1,0 +1,12 @@
+export var browserId2Name = {
+    0: 'Brave',
+    1: 'Chromium',
+    2: 'Chrome',
+    3: 'FireFox',
+    4: 'Safari',
+    5: 'Opera',
+    6: 'Edge',
+    7: 'Edge-Chromium',
+    8: 'IE',
+    9: 'IE8'
+}